Configure Rails for Docker, Create Dockerfile
Resolves:
required that you provide a Dockerfile in the
root directory of the repository.

This Dockerfile must build and run your
project out of the box with no extra steps.

In other words, we should be able to docker
run <image> and have your challenge up and
running.

Attempted to make it simpler to run in one step
through sh script

Only 1 line should get you started:
bin/docker setup